Extract terminal discourse, focus, and particle scope
